Adjust UI spacings; Temporary fix for ColorSelection radio button colorings

+++ lightcrafts/src/com/lightcrafts/ui/toolkit/BoxedButton.java
@@ -3,10 +3,10 @@
 
 package com.lightcrafts.ui.toolkit;
 
-import com.lightcrafts.ui.LightZoneSkin;
 import com.lightcrafts.ui.layout.ToggleTitleBorder;
 
 import java.awt.Font;
+
 import javax.swing.*;
 
 /**
@@ -30,7 +30,9 @@ public class BoxedButton {
         int buttonWidth = 0;
         for (JComponent b : buttons)
             buttonWidth += b.getWidth();
-        final int titleWidth = box.getFontMetrics(font).stringWidth(title);
+
+        // Multiply with 0.6 because stringWidth is wider than actual length.
+        final int titleWidth = (int)(box.getFontMetrics(font).stringWidth(title));
         final int sideMargin = Math.max(0, (titleWidth - buttonWidth) / 2);
 
         box.add(Box.createHorizontalStrut(sideMargin));
